Mise à jour programme Matlab

+++ JCalc.m
@@ -5,7 +5,7 @@ function [ Jo ] = JCalc( nbPixel, G,gradIm )
 for i =1:nbPixel
     j = 2*i;
     a = (gradIm(j-1:j, :))'
-    b = uint8(G(j-1:j,:))
+    b = G(j-1:j,:)
     Jo(i,:) = a*b;
     
 end
